pyproject.toml
[project]
name = "security-intent-auditor"
version = "0.1.0"
description = "AI-powered security auditor based on ISC2 CC principles."
authors = [
    {name = "Carlota de la Vega",email = "carlotadlavega@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "colorama (>=0.4.6,<0.5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
